add: ability to hide test users in actual reports also
